requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
python-dotenv>=1.0.1
openai>=1.40.0
python-telegram-bot==21.4
PySocks>=1.7.1
cryptography>=42.0.8 